dev: Renaming and reorganization after review
