+++ app/clients/model/_albertmodelclient.py
@@ -2,7 +2,6 @@ from urllib.parse import urljoin
 
 from openai import AsyncOpenAI
 import requests
-from httpx import AsyncClient
 
 from app.clients.model._basemodelclient import BaseModelClient
 from app.utils.variables import (
@@ -13,10 +12,7 @@ from app.utils.variables import (
     ENDPOINT__MODELS,
     ENDPOINT__OCR,
     ENDPOINT__RERANK,
-    ENDPOINT__METRICS,
 )
-
-from app.schemas.modelclientmetrics import ModelClientMetrics
 
 
 class AlbertModelClient(AsyncOpenAI, BaseModelClient):
@@ -28,7 +24,6 @@ class AlbertModelClient(AsyncOpenAI, BaseModelClient):
         ENDPOINT__MODELS: "/v1/models",
         ENDPOINT__OCR: "/v1/chat/completions",
         ENDPOINT__RERANK: "/v1/rerank",
-        ENDPOINT__METRICS: None,
     }
 
     def __init__(self, model: str, api_url: str, api_key: str, timeout: int, *args, **kwargs) -> None:
@@ -65,16 +60,3 @@ class AlbertModelClient(AsyncOpenAI, BaseModelClient):
             self.vector_size = len(response.json()["data"][0]["embedding"])
         else:
             self.vector_size = None
-
-    async def get_server_metrics(self) -> ModelClientMetrics | None:
-        headers = {"Authorization": f"Bearer {self.api_key}"} if self.api_key else None
-
-        async with AsyncClient(base_url=self.api_url, headers=headers, timeout=self.timeout) as http_client:
-            response = await http_client.get(self.ENDPOINT_TABLE[ENDPOINT__METRICS])
-            response.raise_for_status()
-
-            metrics = response.json()
-
-            # TODO: Custom mapping of metrics to an harmonized format
-
-            return metrics
