+++ app/clients/model/_openaimodelclient.py
@@ -12,7 +12,6 @@ from app.utils.variables import (
     ENDPOINT__MODELS,
     ENDPOINT__OCR,
     ENDPOINT__RERANK,
-    ENDPOINT__METRICS,
 )
 
 
@@ -25,7 +24,6 @@ class OpenaiModelClient(AsyncOpenAI, BaseModelClient):
         ENDPOINT__MODELS: "/v1/models",
         ENDPOINT__OCR: "/v1/chat/completions",
         ENDPOINT__RERANK: None,
-        ENDPOINT__METRICS: None,
     }
 
     def __init__(self, model: str, api_url: str, api_key: str, timeout: int, *args, **kwargs) -> None:
