+++ app/clients/model/_teimodelclient.py
@@ -16,7 +16,6 @@ from app.utils.variables import (
     ENDPOINT__MODELS,
     ENDPOINT__OCR,
     ENDPOINT__RERANK,
-    ENDPOINT__METRICS,
 )
 
 
@@ -29,7 +28,6 @@ class TeiModelClient(AsyncOpenAI, BaseModelClient):
         ENDPOINT__MODELS: "/info",
         ENDPOINT__OCR: None,
         ENDPOINT__RERANK: "/rerank",
-        ENDPOINT__METRICS: None,
     }
 
     def __init__(self, model: str, api_url: str, api_key: str, timeout: int, *args, **kwargs) -> None:
