+++ app/helpers/_metricstracker.py
@@ -8,10 +8,10 @@ class MetricsTracker:
         self.refresh_rate = refresh_rate
         self.refresh_count_per_window = refresh_count_per_window
 
-    async def monitor(self) -> None:
+    async def start_monitoring(self) -> None:
         # TODO: To implement with calls to get_server_metrics
         pass
 
-    async def stop_monitor(self) -> None:
+    async def stop_monitoring(self) -> None:
         # TODO: To implement
         pass
