+++ app/helpers/message_producer/__init__.py
@@ -1,0 +1,5 @@
+from .rpc_client import RPCClient
+
+__all__ = [
+    "RPCClient",
+]
