+++ app/helpers/models/routers/strategies/leastbusymodelclientselectionstrategy.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from ._basemodelclientselectionstrategy import BaseModelClientSelectionStrategy
+from app.helpers.strategies._basemodelclientselectionstrategy import BaseModelClientSelectionStrategy
 
 
 class LeastBusyModelClientSelectionStrategy(BaseModelClientSelectionStrategy):
