+++ app/utils/variables.py
@@ -18,7 +18,6 @@ ENDPOINT__SEARCH = "/search"
 ENDPOINT__TOKENS = "/tokens"
 ENDPOINT__USERS = "/users"
 ENDPOINT__USERS_ME = "/users/me"
-ENDPOINT__METRICS = "/metrics"
 
 ENDPOINTS = [value for name, value in locals().items() if name.startswith("ENDPOINT__")]
